Extract mode list fetch descriptors
Move right-pane list fetch descriptor logic into model/mode_fetch.go so model_fetch.go stays focused on fetch orchestration and concrete commands. Add a Flow-mode characterization test for selected repo, request tagging, and stale result handling.
