+++ AGENTS.md
@@ -23,7 +23,7 @@ gofmt -l .          # formatting check used by CI
 - `planstore/` persists saved plans beside sessions under the shared agent-artifact root (`<root>/plans/<plan-id>/meta.json` + `plan.md`, dirs `0700`/files `0600`, atomic writes). `Store` exposes `Save` (validates title/content/status, generates timestamp+slug IDs with collision suffixes, merges metadata on update), `List` (sorted by `updated_at` desc, exact cleaned `RepoPath` filter), `ReadPlan`, and `SetPhase`. Store roots must be absolute; an injectable clock keeps ID generation testable.
 - `flowstore/` persists task-centric Flow records beside sessions and plans under `<root>/flows/<flow-id>/meta.json` (dirs `0700`/files `0600`, atomic writes). `Store.Create` requires title, instructions, and an absolute repo path, generates timestamp+slug IDs, seeds the default phase graph (plan, plan review, implementation, review loop, PR creation, autoreview, merge), and derives status from phases/merge metadata. Plan Review gates Implementation: only `approved`/`approved_with_concerns` completed outcomes, or skipped-with-notes, make Implementation ready. `AddChildPhase` creates or updates ordered child phases under Implementation; those children gate review loop and PR creation until completed or skipped with notes. `SetPR` validates GitHub PR metadata, URL shape, positive PR number, base/head branches, and head-vs-Flow branch consistency; Autoreview stays pending after PR Creation until this PR target metadata is recorded. `SetMerge` records explicit merge outcomes: merged requires existing PR metadata, a completed Merge phase, a merge commit, and a merge timestamp, while blocked requires a blocked Merge phase with notes. `SetPlanLink` records linked saved-plan artifacts with locking and absolute plan-path validation, while `List` filters by exact cleaned `RepoPath` and sorts by `updated_at` desc.
 - `internal/artifacts/` owns the shared filesystem mechanics for sessions, plans, and Flows: default artifact-root resolution, absolute-root checks, `0700` collection/record directory setup, `0600` atomic writes from bytes/readers/generator callbacks, safe single-segment IDs, slugging, and timestamp+slug collision allocation. Keep store-specific JSON schemas, filtering, locking, transcript normalization, and phase semantics in the owning store packages.
-- `model/` owns Bubble Tea state, key handling, fuzzy filtering, modal prompts, pager launches, confirmations, and async result handling (split across `model.go`, `model_keys.go`, `model_fetch.go`, `model_messages.go`, `model_filter.go`). New Flow start orchestration is grouped behind the `FlowStarter` in `model/flow_start.go`, which creates the Flow, worktree, start metadata, bootstrap recovery, and plan-phase launch context before the model launches the agent. Each list (repos, branches, stashes, worktrees, commits, reflog, sessions, plans, flows) is a generic value-type `pane.Pane[T]` from `model/pane/` that owns its own filtering, selection, and scroll, and prompt state (confirmations, input prompts, agent selection) is consolidated into a typed `modal.Modal` from `model/modal/`. Read-only content views (diffs, transcripts, summaries, plan bodies) launch `less -R` through `actions.PageText` with stale-result protection via an active view request id. The worktrees pane (mode 1) has separate inline worktree-session state: `x` fetches sessions scoped by repo and selected worktree, arrows select inline sessions while open, `enter` resumes the inline selected session, and context changes close or ignore stale inline results without populating the mode 6 sessions pane. The sessions pane is right-pane mode 6; `enter` pages the normalized transcript, `r` resumes the stored provider session, `s` pages the summary, and `y` copies the raw provider session ID. The plans pane is right-pane mode 7; `enter` expands/collapses phase rows, `o` pages the Markdown, `i` edits launch instructions before launching the selected agent, and `y` copies the plan Markdown path. The production TUI starts in the flows pane (mode 8), while the generic model constructor still defaults to worktrees unless `Options.StartupMode` is set. The flows pane shows linked plan and current phase state, `n` creates a new Flow, `x` expands/collapses phase detail rows for the selected Flow, `o` pages the linked plan body or reports when no plan is linked, and `a` launches the selected ready Flow phase while recording launch metadata.
+- `model/` owns Bubble Tea state, key handling, fuzzy filtering, modal prompts, pager launches, confirmations, and async result handling (split across `model.go`, `model_keys.go`, `model_fetch.go`, `mode_fetch.go`, `model_messages.go`, `model_filter.go`). New Flow start orchestration is grouped behind the `FlowStarter` in `model/flow_start.go`, which creates the Flow, worktree, start metadata, bootstrap recovery, and plan-phase launch context before the model launches the agent. Each list (repos, branches, stashes, worktrees, commits, reflog, sessions, plans, flows) is a generic value-type `pane.Pane[T]` from `model/pane/` that owns its own filtering, selection, and scroll, and prompt state (confirmations, input prompts, agent selection) is consolidated into a typed `modal.Modal` from `model/modal/`. Read-only content views (diffs, transcripts, summaries, plan bodies) launch `less -R` through `actions.PageText` with stale-result protection via an active view request id. The worktrees pane (mode 1) has separate inline worktree-session state: `x` fetches sessions scoped by repo and selected worktree, arrows select inline sessions while open, `enter` resumes the inline selected session, and context changes close or ignore stale inline results without populating the mode 6 sessions pane. The sessions pane is right-pane mode 6; `enter` pages the normalized transcript, `r` resumes the stored provider session, `s` pages the summary, and `y` copies the raw provider session ID. The plans pane is right-pane mode 7; `enter` expands/collapses phase rows, `o` pages the Markdown, `i` edits launch instructions before launching the selected agent, and `y` copies the plan Markdown path. The production TUI starts in the flows pane (mode 8), while the generic model constructor still defaults to worktrees unless `Options.StartupMode` is set. The flows pane shows linked plan and current phase state, `n` creates a new Flow, `x` expands/collapses phase detail rows for the selected Flow, `o` pages the linked plan body or reports when no plan is linked, and `a` launches the selected ready Flow phase while recording launch metadata.
 - `ui/` is stateless rendering with lipgloss, driven by a `RenderParams` snapshot. Flow rendering includes recovery labels for partial states: `recover-worktree`, `await-session`, `session-mismatch`, and `missing-pr`.
 - `internal/version/` holds version/commit/date injected via `-ldflags` at build time.
